# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Interview documentation",
        "description": "Interview documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/interviews": {
            "get": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Get all interviews",
                "operationId": "a739402bae882f3a7ea127d27664ade8",
                "responses": {
                    "200": {
                        "description": "List of all interviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "javascript-basics"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "JavaScript Basics Interview"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Basic JavaScript interview questions and answers"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "level": {
                                                "type": "string",
                                                "example": "Beginner"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Create a new interview",
                "operationId": "d85cf3e25f2f8f6b358f3cc08713aed8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "slug",
                                    "title",
                                    "description",
                                    "category_id",
                                    "level"
                                ],
                                "properties": {
                                    "slug": {
                                        "type": "string",
                                        "example": "javascript-basics"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "JavaScript Basics Interview"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Basic JavaScript interview questions and answers"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Beginner"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Interview created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "javascript-basics"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "JavaScript Basics Interview"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Basic JavaScript interview questions and answers"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "level": {
                                            "type": "string",
                                            "example": "Beginner"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "slug": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The slug has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/interviews/{id}": {
            "put": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Update an existing interview",
                "operationId": "56a7830c19de17f722384949f587e567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Interview ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "slug": {
                                        "type": "string",
                                        "example": "updated-javascript-basics"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Updated JavaScript Basics Interview"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated JavaScript interview questions and answers"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Intermediate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interview updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "updated-javascript-basics"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Updated JavaScript Basics Interview"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Updated JavaScript interview questions and answers"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "level": {
                                            "type": "string",
                                            "example": "Intermediate"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Interview not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Interview] 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "slug": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The slug has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Interviews"
                ],
                "summary": "Delete an interview",
                "operationId": "70bc2a317c7a3b907f9faba6fad82fb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Interview ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Interview deleted successfully"
                    },
                    "404": {
                        "description": "Interview not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Interview] 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}